+++ userscript/vite.config.ts
@@ -46,7 +46,7 @@ export default defineConfig({
       entry: 'src/main.ts',
       userscript: {
         icon: 'https://learnablemeta.com/favicon.png',
-        version: '0.91',
+        version: '0.92',
         namespace: 'geometa',
         name: 'GeoGuessr Learnable Meta',
         description: 'UserScript for GeoGuessr Learnable Meta maps',